lightfx: Move Color and Frame into separate crate
