+++ Cargo.toml
@@ -3,7 +3,7 @@ members = [
     "light-client",
     "configurator",
     "animator",
-    "animation-model",
+    "lightfx",
     "visualiser",
     "webapi",
     "webui",
