--- animation-model/src/lib.rs
@@ -1,3 +1,0 @@
-pub mod parameter_schema;
-
-pub use parameter_schema as schema;
