+++ animator/src/animations/check.rs
@@ -1,6 +1,5 @@
 use super::Animation;
 use log::debug;
-use rustmas_light_client as client;
 
 pub struct Check {
     points_count: usize,
@@ -15,9 +14,9 @@ impl Check {
 }
 
 impl Animation for Check {
-    fn frame(&mut self, time: f64) -> client::Frame {
+    fn frame(&mut self, time: f64) -> lightfx::Frame {
         let index = ((time * 8.0) % self.points_count as f64) as usize;
         debug!("Checking light #{}", index);
-        client::Frame::new_black(self.points_count).with_pixel(index, client::Color::white())
+        lightfx::Frame::new_black(self.points_count).with_pixel(index, lightfx::Color::white())
     }
 }
