+++ animator/src/animations/rainbow_cable.rs
@@ -1,5 +1,4 @@
 use super::Animation;
-use rustmas_light_client as client;
 
 pub struct RainbowCable {
     points_count: usize,
@@ -14,10 +13,12 @@ impl RainbowCable {
 }
 
 impl Animation for RainbowCable {
-    fn frame(&mut self, time: f64) -> client::Frame {
+    fn frame(&mut self, time: f64) -> lightfx::Frame {
         (0..self.points_count)
             .into_iter()
-            .map(|i| client::Color::hsv(i as f64 / self.points_count as f64 * 4.0 + time, 1.0, 0.5))
+            .map(|i| {
+                lightfx::Color::hsv(i as f64 / self.points_count as f64 * 4.0 + time, 1.0, 0.5)
+            })
             .into()
     }
 }
