+++ animator/src/animations/rainbow_spiral.rs
@@ -1,5 +1,4 @@
 use super::{utils, Animation};
-use rustmas_light_client as client;
 
 pub struct RainbowSpiral {
     points_polar: Vec<(f64, f64, f64)>,
@@ -14,11 +13,11 @@ impl RainbowSpiral {
 }
 
 impl Animation for RainbowSpiral {
-    fn frame(&mut self, time: f64) -> client::Frame {
+    fn frame(&mut self, time: f64) -> lightfx::Frame {
         self.points_polar
             .iter()
             .map(|(_, a, h)| {
-                client::Color::hsv(a / (std::f64::consts::PI * 2.0) + h + time / 2.0, 1.0, 0.5)
+                lightfx::Color::hsv(a / (std::f64::consts::PI * 2.0) + h + time / 2.0, 1.0, 0.5)
             })
             .into()
     }
