+++ configurator/Cargo.toml
@@ -5,12 +5,13 @@ edition = "2021"
 
 [dependencies]
 rustmas-light-client = { path = "../light-client", default-features = false }
-opencv = "0.70"
+lightfx = { path = "../lightfx" }
 
 csv = "1.1.6"
 clap = { version = "4.0.8", features = ["derive"] }
 indicatif = "0.17.1"
 itertools = "0.10.5"
+opencv = "0.70"
 tokio = { version = "1", features = ["full"] }
 
 log = "0.4.17"
