+++ lightfx/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "rustmas-animation-model"
+name = "lightfx"
 version = "0.1.0"
 edition = "2021"
