+++ lightfx/src/lib.rs
@@ -1,0 +1,7 @@
+mod color;
+mod frame;
+pub mod parameter_schema;
+
+pub use color::Color;
+pub use frame::Frame;
+pub use parameter_schema as schema;
