+++ webui/src/controls/color_control.rs
@@ -1,39 +1,8 @@
-use rustmas_animation_model::parameter_schema::{Parameter, ParameterValue};
-use serde_json::json;
+use lightfx::parameter_schema::{Parameter, ParameterValue};
 use web_sys::HtmlInputElement;
 use yew::{html, Component, Context, Html, NodeRef};
 
 use super::ParameterControlProps;
-
-fn hex_from_str(code: &str) -> Option<serde_json::Value> {
-    let code = code.trim_start_matches(|c| c == '#');
-    let (r, g, b) = match code.len() {
-        6 => match u32::from_str_radix(code, 16) {
-            Ok(x) => ((x & 0xFF0000) >> 16, (x & 0x00FF00) >> 8, x & 0x0000FF),
-            Err(_) => return None,
-        },
-        3 => match u32::from_str_radix(code, 16) {
-            Ok(x) => (
-                ((x & 0xF00) >> 8) * 0x11,
-                ((x & 0x0F0) >> 4) * 0x11,
-                (x & 0x00F) * 0x11,
-            ),
-            Err(_) => return None,
-        },
-        _ => return None,
-    };
-
-    Some(json!({"r": r, "g": g, "b": b}))
-}
-
-fn hex_from_value(value: &serde_json::Value) -> String {
-    format!(
-        "#{:02x}{:02x}{:02x}",
-        value.get("r").unwrap().as_i64().unwrap(),
-        value.get("g").unwrap().as_i64().unwrap(),
-        value.get("b").unwrap().as_i64().unwrap()
-    )
-}
 
 #[derive(Default)]
 pub struct ColorInput {}
@@ -63,8 +32,12 @@ impl Component for ColorParameterControl {
                     self.node_ref.cast::<HtmlInputElement>(),
                     self.hidden_ref.cast::<HtmlInputElement>(),
                 ) {
-                    (Some(node), Some(hidden)) => hidden
-                        .set_value(&serde_json::to_string(&hex_from_str(&node.value())).unwrap()),
+                    (Some(node), Some(hidden)) => hidden.set_value(
+                        &serde_json::to_string(
+                            &lightfx::Color::from_hex_str(&node.value()).unwrap(),
+                        )
+                        .unwrap(),
+                    ),
                     _ => (),
                 };
                 false
@@ -82,8 +55,9 @@ impl Component for ColorParameterControl {
             let value = ctx
                 .props()
                 .value
-                .as_ref()
-                .map(hex_from_value)
+                .clone()
+                .and_then(|v| serde_json::from_value::<lightfx::Color>(v).ok())
+                .map(|v| lightfx::Color::to_hex_string(&v))
                 .unwrap_or("#000000".to_owned());
             let value_hex = serde_json::to_string(&ctx.props().value).unwrap();
